+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KJ502Q19JWFD31CB19HESYKJ"
+        "01KJ52MNYT3WXW967HDHBDZN0M"
     )
     r"""The unique identifier of the data source"""