+++ packages/orq-rc/RELEASES.md
@@ -5969,3 +5969,13 @@ Based on:
 - [python v4.4.0-rc.29] packages/orq-rc
 ### Releases
 - [PyPI v4.4.0-rc.29] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.29 - packages/orq-rc
+
+## 2026-02-23 11:05:04
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.724.1 (2.836.5) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [python v4.4.0-rc.30] packages/orq-rc
+### Releases
+- [PyPI v4.4.0-rc.30] https://pypi.org/project/orq-ai-sdk/4.4.0-rc.30 - packages/orq-rc